chore: improve range and cache calculation
