+++ src/core/completion/register.ts
@@ -53,7 +53,7 @@ export const registerCompletion = (
   editor.updateOptions({
     inlineSuggest: {
       enabled: true,
-      mode: 'prefix',
+      mode: 'subwordSmart',
     },
   });
