+++ src/types/completion.ts
@@ -262,4 +262,6 @@ export interface CompletionCacheItem {
   completion: string;
   range: EditorRange;
   textBeforeCursor: string;
+  textAfterCursor: string;
+  cachePos: CursorPosition;
 }
