+++ tests/ui/src/app/api/complete/route.ts
@@ -1,8 +1,8 @@
 import {Copilot} from 'monacopilot';
 
-const copilot = new Copilot(process.env.OPENAI_API_KEY!, {
-  provider: 'google',
-  model: 'gemini-1.5-flash',
+const copilot = new Copilot(process.env.GROQ_API_KEY!, {
+  provider: 'groq',
+  model: 'llama-3-70b',
 });
 
 export async function POST(req: Request) {
