+++ tests/ui/src/app/page.tsx
@@ -20,6 +20,7 @@ export default function Home() {
       endpoint: '/api/complete',
       language: 'javascript',
       maxContextLines: 60,
+      trigger: 'onTyping',
     });
 
     // monaco.editor.addEditorAction({
